app/requirements.txt
django==3.2
djangorestframework
drf_yasg
django-filter
psycopg2-binary
gunicorn